# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Users API",
        "version": "0.1"
    },
    "paths": {
        "/api/users": {
            "get": {
                "operationId": "eaedf6aa84fde38a08bf5d87308169f7",
                "responses": {
                    "200": {
                        "description": "Listado de usuarios"
                    }
                }
            }
        },
        "/api/users/create": {
            "get": {
                "operationId": "ce291a12c8da6f715f6f562df3be7466",
                "responses": {
                    "200": {
                        "description": "Datos necesarios para la creación de usuarios"
                    }
                }
            }
        },
        "/api/users/": {
            "post": {
                "description": "Registro de usuario",
                "operationId": "c27272d59ef5175a5459f61bcf048622",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "address",
                                    "phone",
                                    "phone_2",
                                    "postal_code",
                                    "birth_date",
                                    "gender"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "text"
                                    },
                                    "last_name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "text"
                                    },
                                    "address": {
                                        "type": "text"
                                    },
                                    "phone": {
                                        "type": "text"
                                    },
                                    "phone_2": {
                                        "type": "text"
                                    },
                                    "postal_code": {
                                        "type": "text"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario registrado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}/": {
            "put": {
                "description": "Registro de usuario",
                "operationId": "25accfcdaf187595c580d89c1ea2fcf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UID del usuario a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "address",
                                    "phone",
                                    "phone_2",
                                    "postal_code",
                                    "birth_date",
                                    "gender"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "text"
                                    },
                                    "address": {
                                        "type": "text"
                                    },
                                    "phone": {
                                        "type": "text"
                                    },
                                    "phone_2": {
                                        "type": "text"
                                    },
                                    "postal_code": {
                                        "type": "text"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario registrado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Eliminación de usuario",
                "operationId": "bfc4022b178721be6d9db06bfd47a9ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UID del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario registrado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}